unify common plug operations
